refactor: 도메인만 별도 패키지로 분리
