+++ src/main/java/goormthonuniv/kengdori/backend/place/DTO/BoundedPlaceResponseDTO.java
@@ -1,8 +1,8 @@
 package goormthonuniv.kengdori.backend.place.DTO;
 
 import goormthonuniv.kengdori.backend.hashtag.DTO.HashtagInfoDTO;
-import goormthonuniv.kengdori.backend.hashtag.domain.PlaceHashtag;
-import goormthonuniv.kengdori.backend.place.domain.Place;
+import goormthonuniv.kengdori.backend.domain.PlaceHashtag;
+import goormthonuniv.kengdori.backend.domain.Place;
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
