+++ src/main/java/goormthonuniv/kengdori/backend/place/controller/PlaceController.java
@@ -4,16 +4,13 @@ import goormthonuniv.kengdori.backend.global.JWT.JwtUtil;
 import goormthonuniv.kengdori.backend.place.DTO.BoundedPlaceListResponseDTO;
 import goormthonuniv.kengdori.backend.place.DTO.BoundedPlaceResponseDTO;
 import goormthonuniv.kengdori.backend.place.service.PlaceService;
-import goormthonuniv.kengdori.backend.user.domain.User;
+import goormthonuniv.kengdori.backend.domain.User;
 import goormthonuniv.kengdori.backend.user.service.UserServiceImpl;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
