+++ src/main/java/goormthonuniv/kengdori/backend/review/DTO/ReviewListResponseDTO.java
@@ -1,7 +1,7 @@
 package goormthonuniv.kengdori.backend.review.DTO;
 
-import goormthonuniv.kengdori.backend.hashtag.domain.PlaceHashtag;
-import goormthonuniv.kengdori.backend.review.domain.Review;
+import goormthonuniv.kengdori.backend.domain.PlaceHashtag;
+import goormthonuniv.kengdori.backend.domain.Review;
 import goormthonuniv.kengdori.backend.hashtag.DTO.HashtagInfoDTO;
 import lombok.Getter;
