+++ src/main/java/goormthonuniv/kengdori/backend/review/service/ReviewService.java
@@ -1,6 +1,6 @@
 package goormthonuniv.kengdori.backend.review.service;
 
-import goormthonuniv.kengdori.backend.user.domain.User;
+import goormthonuniv.kengdori.backend.domain.User;
 import goormthonuniv.kengdori.backend.place.DTO.VisitedPlaceResponseDTO;
 import goormthonuniv.kengdori.backend.review.DTO.*;
 import org.springframework.data.domain.Page;
